Trying to use upstream sensor plugin
This PR removes the science sensor system plugin and tries to use the upstream plugin.

+++ lrauv_ignition_plugins/src/TethysCommPlugin.cc
@@ -487,7 +487,7 @@ void TethysCommPlugin::BuoyancyStateCallback(
 }
 
 void TethysCommPlugin::SalinityCallback(
-  const gz::msgs::Float &_msg)
+  const gz::msgs::Double &_msg)
 {
   this->latestSalinity = _msg.data();
 }
@@ -508,7 +508,7 @@ void TethysCommPlugin::BatteryCallback(
 }
 
 void TethysCommPlugin::ChlorophyllCallback(
-  const gz::msgs::Float &_msg)
+  const gz::msgs::Double &_msg)
 {
   this->latestChlorophyll = _msg.data();
 }
